refactors and optimizations
